requirements.txt
Flask==3.0.0
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
playwright==1.38.0

googleapis-common-protos==1.60.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
ipdb==0.13.13
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
python-dotenv==1.0.0
colorama==0.4.6
